[ENH]: bundle CLI with JS package
