+++ clients/chroma-cli/.gitignore
@@ -1,0 +1,8 @@
+target
+index.node
+**/node_modules
+**/.DS_Store
+npm-debug.log*
+lib
+cargo.log
+cross.log
