+++ clients/chroma-cli/bin-entrypoint.cjs
@@ -1,0 +1,3 @@
+#!/usr/bin/env node
+const { cli } = require("./lib/index.cjs");
+cli(...process.argv.slice(2));
