+++ clients/chroma-cli/platforms/darwin-arm64/README.md
@@ -1,0 +1,3 @@
+# `@chroma-cli/darwin-arm64`
+
+Prebuilt binary package for `chroma-cli` on `darwin-arm64`.
