+++ clients/chroma-cli/platforms/darwin-arm64/package.json
@@ -1,0 +1,25 @@
+{
+  "name": "@chroma-cli/darwin-arm64",
+  "description": "Prebuilt binary package for `chroma-cli` on `darwin-arm64`.",
+  "version": "0.1.0",
+  "os": [
+    "darwin"
+  ],
+  "cpu": [
+    "arm64"
+  ],
+  "main": "index.node",
+  "files": [
+    "index.node"
+  ],
+  "neon": {
+    "type": "binary",
+    "rust": "aarch64-apple-darwin",
+    "node": "darwin-arm64",
+    "os": "darwin",
+    "arch": "arm64",
+    "abi": null
+  },
+  "author": "",
+  "license": "ISC"
+}
