+++ clients/chroma-cli/platforms/darwin-x64/README.md
@@ -1,0 +1,3 @@
+# `@chroma-cli/darwin-x64`
+
+Prebuilt binary package for `chroma-cli` on `darwin-x64`.
