+++ clients/chroma-cli/platforms/linux-arm64-gnu/README.md
@@ -1,0 +1,3 @@
+# `@chroma-cli/linux-arm64-gnu`
+
+Prebuilt binary package for `chroma-cli` on `linux-arm64-gnu`.
