+++ clients/chroma-cli/platforms/linux-arm64-gnu/package.json
@@ -1,0 +1,25 @@
+{
+  "name": "@chroma-cli/linux-arm64-gnu",
+  "description": "Prebuilt binary package for `chroma-cli` on `linux-arm64-gnu`.",
+  "version": "0.1.0",
+  "os": [
+    "linux"
+  ],
+  "cpu": [
+    "arm64"
+  ],
+  "main": "index.node",
+  "files": [
+    "index.node"
+  ],
+  "neon": {
+    "type": "binary",
+    "rust": "aarch64-unknown-linux-gnu",
+    "node": "linux-arm64-gnu",
+    "os": "linux",
+    "arch": "arm64",
+    "abi": "gnu"
+  },
+  "author": "",
+  "license": "ISC"
+}
