+++ clients/chroma-cli/platforms/linux-x64-gnu/README.md
@@ -1,0 +1,3 @@
+# `@chroma-cli/linux-x64-gnu`
+
+Prebuilt binary package for `chroma-cli` on `linux-x64-gnu`.
