+++ clients/chroma-cli/platforms/win32-x64-msvc/README.md
@@ -1,0 +1,3 @@
+# `@chroma-cli/win32-x64-msvc`
+
+Prebuilt binary package for `chroma-cli` on `win32-x64-msvc`.
