+++ clients/chroma-cli/src/index.mts
@@ -1,0 +1,3 @@
+// This module is the ESM entry point for the library.
+
+export * from './index.cjs';
