+++ clients/chroma-cli/tsconfig.json
@@ -1,0 +1,9 @@
+{
+  "extends": "@tsconfig/node20/tsconfig.json",
+  "compilerOptions": {
+    "module": "node16",
+    "declaration": true,
+    "outDir": "lib",
+  },
+  "exclude": ["lib"]
+}
