+++ rust/cli/Cargo.toml
@@ -1,10 +1,10 @@
 [package]
-name = "cli"
+name = "chroma-cli"
 version = "0.1.0"
 edition = "2021"
 
 [dependencies]
-clap = { version = "4.5.28", features = ["derive"] }
+clap = { workspace = true }
 webbrowser = "1.0.3"
 tokio = { workspace = true }
