+++ rust/cli/src/main.rs
@@ -1,75 +1,7 @@
-use chroma_frontend::{config::FrontendServerConfig, frontend_service_entrypoint_with_config};
-use clap::{Parser, Subcommand};
-use std::sync::Arc;
-
-const LOGO: &str = "
-                \x1b[38;5;069m(((((((((    \x1b[38;5;203m(((((\x1b[38;5;220m####
-             \x1b[38;5;069m(((((((((((((\x1b[38;5;203m(((((((((\x1b[38;5;220m#########
-           \x1b[38;5;069m(((((((((((((\x1b[38;5;203m(((((((((((\x1b[38;5;220m###########
-         \x1b[38;5;069m((((((((((((((\x1b[38;5;203m((((((((((((\x1b[38;5;220m############
-        \x1b[38;5;069m(((((((((((((\x1b[38;5;203m((((((((((((((\x1b[38;5;220m#############
-        \x1b[38;5;069m(((((((((((((\x1b[38;5;203m((((((((((((((\x1b[38;5;220m#############
-         \x1b[38;5;069m((((((((((((\x1b[38;5;203m(((((((((((((\x1b[38;5;220m##############
-         \x1b[38;5;069m((((((((((((\x1b[38;5;203m((((((((((((\x1b[38;5;220m##############
-           \x1b[38;5;069m((((((((((\x1b[38;5;203m(((((((((((\x1b[38;5;220m#############
-             \x1b[38;5;069m((((((((\x1b[38;5;203m((((((((\x1b[38;5;220m##############
-                \x1b[38;5;069m(((((\x1b[38;5;203m((((    \x1b[38;5;220m#########\x1b[0m
-";
-
-#[derive(Parser, Debug)]
-struct RunArgs {
-    #[clap(name = "config", default_value = None)]
-    config: Option<String>,
-}
-
-#[derive(Subcommand, Debug)]
-enum Command {
-    Docs,
-    Run(RunArgs),
-    Support,
-}
-
-#[derive(Parser, Debug)]
-#[command(name = "chroma")]
-#[command(version = "0.0.1")]
-#[command(about = "A CLI for Chroma", long_about = None)]
-struct Cli {
-    #[command(subcommand)]
-    command: Command,
-}
-
-fn run(args: RunArgs) {
-    println!("{}", LOGO);
-
-    let config = match &args.config {
-        Some(path) => FrontendServerConfig::load_from_path(path),
-        None => FrontendServerConfig::single_node_default(),
-    };
-
-    let runtime = tokio::runtime::Runtime::new().expect("Failed to start Chroma");
-    runtime.block_on(async {
-        frontend_service_entrypoint_with_config(Arc::new(()), Arc::new(()), (), config).await;
-    });
-}
+use chroma_cli::{run_command_from_args, Cli};
+use clap::Parser;
 
 fn main() {
     let cli = Cli::parse();
-
-    match cli.command {
-        Command::Docs => {
-            let url = "https://docs.trychroma.com";
-            if webbrowser::open(url).is_err() {
-                eprintln!("Error: Failed to open the browser. Visit {}.", url);
-            }
-        }
-        Command::Run(args) => {
-            run(args);
-        }
-        Command::Support => {
-            let url = "https://discord.gg/MMeYNTmh3x";
-            if webbrowser::open(url).is_err() {
-                eprintln!("Error: Failed to open the browser. Visit {}.", url);
-            }
-        }
-    }
+    run_command_from_args(cli);
 }
